+++ src/commands/index.ts
@@ -1,16 +1,21 @@
-import { log } from 'console';
 import * as vscode from 'vscode';
 import { FSItem } from '../FSItem';
 import { TreeDataProvider } from '../providers/TreeDataProvider';
-import { getCrudCommands } from './crud';
+import { log } from '../utils';
+import { getCrudCommands, UndoAction } from './crud';
 import { getGeneralCommands } from './general';
 import { getRevealCommands } from './reveal';
 import { getToggleCommands } from './toggle';
 
-export function registerCommands(context: vscode.ExtensionContext, treeView: vscode.TreeView<FSItem>, provider: TreeDataProvider) {
+export function registerCommands(
+  context: vscode.ExtensionContext,
+  treeView: vscode.TreeView<FSItem>,
+  provider: TreeDataProvider,
+  undoState: { action: UndoAction | null },
+) {
   const commands = {
     ...getGeneralCommands(treeView, provider),
-    ...getCrudCommands(treeView, provider),
+    ...getCrudCommands(treeView, provider, undoState),
     ...getRevealCommands(treeView, provider),
     ...getToggleCommands(treeView),
   };